fix parsing of SysInfoExtended for nano 6g
The attached patch fixes an issue when parsing SysInfoExtended on iPod
nano gen. 6 devices:

Apparently they have now two tags for the following lists:
AlbumArt (and AlbumArt2)
ImageSpecifications (and ImageSpecifications2)
ChapterImageSpecs (and ChapterImageSpecs2)

e.g.:
---------------------
[...]
<key>ChapterImageSpecs</key>
<array>
</array>
<key>ChapterImageSpecs2</key>
<array>
<dict>
<key>FormatId</key>
<integer>1073</integer>
<key>RenderWidth</key>
<integer>240</integer>
<key>RenderHeight</key>
[...]
---------------------

The <array> following the first tags are always empty and the <array>
following the 2nd type of tags (with the "2" at the end) contains the
full list of the respective items.

I have checked the following:
- SysInfoExtended is correctly written (I have verified that it matches
the xml data which is retrieved via raw SCSI inquiry commands as well as
via the vendor specific USB requests), so it looks like that they really
added these new tags.

- I have updated my iPod to the newest firmware 1.1 and it's still the same