don't crash on NULL in itdb_sysinfo_properties_free
If the passed in pointer is NULL, itdb_sysinfo_properties_free
will attempt to dereference it which will cause a crash. This crash
was triggered by the udev callout code if
itdb_sysinfo_extended_parse_from_xml returned NULL. This happened
for example with an iPod Shuffle 4G which returned "" in answer
to the USB inquiry command.