udev: rework udev rule to handle Shuffle 4g
Some iPod Shuffle (4G) don't have a partition, only a filesystem on the
block device. The current rule won't match for such iPods because it wants
a partition. Given that some Nano5g have been seen as reporting a
filesystem on the block device while they have a partition, we don't want
to run the rule unconditionally on any block device or partition. So we add
an additional rule which matches devices with a partition, but with a check
for ID_PART_TABLE_TYPE. if it's empty, it means we only have a block
device, in which case we can run the callout.
This bug was fixed thanks to eshat's investigations on IRC.