fix (again) creation of non-existing Device dirs
Recent iPods (Shuffle, iOS devices, ...) don't have a Device/
subdirectory in their control dir, so we need to create it
ourselves before trying to write a SysInfoExtended file to it.
However, due to a "typo", we were creating
iPod_Control/iTunes/Device/ instead of iPod_Control/Device. This
commit fixes this, even though in the long run we should stop
using the iphone-set-info callout for iOS devices and read the
needed information directly from the device.